requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
python-multipart>=0.0.12
aiofiles>=24.1.0
pydantic>=2.9.0
eventlet>=0.36.1
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
bcrypt>=4.0.1

